Functions separated into the modules
